adopt epicflare-style user seeding workflow
