+++ shared/hex.ts
@@ -1,0 +1,5 @@
+export function toHex(bytes: Uint8Array) {
+	return Array.from(bytes)
+		.map((value) => value.toString(16).padStart(2, '0'))
+		.join('')
+}
